requirements.txt
pandas
plotly
google-cloud-storage
tabulate